Docs - messages, future feats, technologies, ERD

+++ app_home/views.py
@@ -13,6 +13,7 @@ from .forms import (
     ProjectLangSelectFilterForm,
 )
 
+
 class HomeView(TemplateView):
     """
     View for the home page.
@@ -153,10 +154,7 @@ class ProjectDetailView(LoginRequiredMixin, TemplateView):
         project_pk = kwargs["pk"]
         project_to_get = get_object_or_404(Project, pk=project_pk)
 
-        context = {
-            "user": request.user,
-            "project": project_to_get
-            }
+        context = {"user": request.user, "project": project_to_get}
 
         return render(request, "app_home/project_detail_view.html", context)
 
@@ -278,8 +276,13 @@ class DeleteProjectView(LoginRequiredMixin, TemplateView):
 
     def get(self, request, *args, **kwargs):
         project = get_object_or_404(Project, pk=kwargs["pk"])
-        project.delete()
-        messages.success(request, "Project successfully deleted!")
+        if project:
+            project.delete()
+            messages.success(request, "Project successfully deleted!")
+        else:
+            messages.error(
+                request, "Project could not be deleted. Please try again."
+            )
         return redirect(reverse("app_home:project-overview"))
